Fix remaining int narrowing in _to_map call site and mp2 loop counters
Address Copilot review comments:
- sparse.cpp: change static_cast<int> to static_cast<size_t> at _to_map
  call site to avoid truncating Eigen::Index for large orbital counts.
- mp2.cpp: widen loop counters from int to Eigen::Index with explicit
  cast of the size_t bound, eliminating signed/unsigned mismatch and
  potential infinite loop when active_space_size exceeds INT_MAX.

+++ cpp/src/qdk/chemistry/data/hamiltonian_containers/sparse.cpp
@@ -93,7 +93,7 @@ SparseHamiltonianContainer::SparseHamiltonianContainer(
           type),
       _one_body_sparse(_to_sparse(one_body_integrals)),
       _two_body_map(_to_map(two_body_integrals,
-                            static_cast<int>(one_body_integrals.rows()))) {}
+                            static_cast<size_t>(one_body_integrals.rows()))) {}
 
 SparseHamiltonianContainer::SparseHamiltonianContainer(
     const Eigen::MatrixXd& one_body_integrals, double core_energy,
